自动化相关代码/atf-server-merge_two_20191113/requirements.txt
aioredis==1.2.0
amqp==2.5.0
appdirs==1.4.3
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
Babel==2.7.0
bcrypt==3.1.6
billiard==3.6.0.0
celery==4.3.0
celery-once==3.0.0
certifi==2019.3.9
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
django-filters==0.2.1
django-guardian==2.0.0
django-simple-history==2.7.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
fabric2==2.4.0
flower==0.9.3
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
invoke==1.2.0
itypes==1.1.0
Jinja2==2.10.1
kombu==4.6.3
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==0.6.1
mysqlclient==1.3.14
nose==1.3.7
olefile==0.46
paramiko==2.4.2
paramiko-expect==0.2.8
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-nmap==0.6.1
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.22.0
ruamel-yaml==0.15.97
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
tornado==5.1.1
Twisted==19.2.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
zope-interface==4.6.0
colorama==0.4.1
colorlog==4.0.2
har2case==0.3.1
jinja2==2.10.3
pyyaml==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
filetype==1.0.5
jsonpath==0.82